添加基金排序worker,修复部分webpack配置
