+++ src/main/workers/index.ts
@@ -1,0 +1,3 @@
+import RequestWorker from './request.worker.ts';
+
+export const requestWorker = new RequestWorker();
