+++ src/main/workers/request.worker.ts
@@ -2,7 +2,13 @@ import got from 'got';
 import Proxy from '../proxy';
 import registerPromiseWorker from 'promise-worker/register';
 
-registerPromiseWorker(async function ({ url, config, proxyConent }: any) {
+interface WorkerRecieveData {
+  url: string;
+  proxyConent?: string;
+  config: any;
+}
+
+registerPromiseWorker(async function ({ url, config, proxyConent }: WorkerRecieveData) {
   try {
     const { httpAgent, httpsAgent } = new Proxy(proxyConent, url);
     const { body, rawBody } = await got(url, {
