+++ src/renderer/index.tsx
@@ -10,7 +10,7 @@ import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import { Provider } from 'react-redux';
 import store from '@/store';
 import App from '@/App';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import 'electron-disable-file-drop';
 import 'dayjs/locale/zh-cn';
 import '@/utils/window';
@@ -23,7 +23,7 @@ echarts.registerMap('china', chinaMap as any);
 
 NP.enableBoundaryChecking(false);
 
-Utils.CheckEnvTool();
+Enhancement.CheckEnvTool();
 
 const { platform } = window.contextModules.process;
