+++ src/renderer/store/listeners/setting.ts
@@ -1,22 +1,22 @@
 import listenerMiddleware from '@/store/listeners';
 import { updateAdjustmentNotificationDateAction, syncSettingAction } from '@/store/features/setting';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import * as CONST from '@/constants';
 
 export default () => {
   listenerMiddleware.startListening({
     actionCreator: syncSettingAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.SYSTEM_SETTING, action.payload);
+      Enhancement.SetStorage(CONST.STORAGE.SYSTEM_SETTING, action.payload);
     },
   });
   listenerMiddleware.startListening({
     actionCreator: updateAdjustmentNotificationDateAction,
     effect: async (action) => {
       if (action.payload) {
-        Utils.SetStorage(CONST.STORAGE.ADJUSTMENT_NOTIFICATION_DATE, action.payload);
+        Enhancement.SetStorage(CONST.STORAGE.ADJUSTMENT_NOTIFICATION_DATE, action.payload);
       } else {
-        Utils.ClearStorage(CONST.STORAGE.ADJUSTMENT_NOTIFICATION_DATE);
+        Enhancement.ClearStorage(CONST.STORAGE.ADJUSTMENT_NOTIFICATION_DATE);
       }
     },
   });
