+++ src/renderer/store/listeners/sort.ts
@@ -1,19 +1,19 @@
 import listenerMiddleware from '@/store/listeners';
 import { syncSortModeAction, syncViewModeAction } from '@/store/features/sort';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import * as CONST from '@/constants';
 
 export default () => {
   listenerMiddleware.startListening({
     actionCreator: syncSortModeAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.SORT_MODE, action.payload);
+      Enhancement.SetStorage(CONST.STORAGE.SORT_MODE, action.payload);
     },
   });
   listenerMiddleware.startListening({
     actionCreator: syncViewModeAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.VIEW_MODE, action.payload);
+      Enhancement.SetStorage(CONST.STORAGE.VIEW_MODE, action.payload);
     },
   });
 };
