+++ src/renderer/store/listeners/stock.ts
@@ -1,13 +1,13 @@
 import listenerMiddleware from '@/store/listeners';
 import { syncStocksConfigAction } from '@/store/features/stock';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import * as CONST from '@/constants';
 
 export default () => {
   listenerMiddleware.startListening({
     actionCreator: syncStocksConfigAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.STOCK_SETTING, action.payload.stockConfig);
+      Enhancement.SetStorage(CONST.STORAGE.STOCK_SETTING, action.payload.stockConfig);
     },
   });
 };
