+++ src/renderer/store/listeners/tabs.ts
@@ -1,13 +1,13 @@
 import listenerMiddleware from '@/store/listeners';
 import { syncTabsActiveKeyAction } from '@/store/features/tabs';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import * as CONST from '@/constants';
 
 export default () => {
   listenerMiddleware.startListening({
     actionCreator: syncTabsActiveKeyAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.TABS_ACTIVE_KEY, action.payload);
+      Enhancement.SetStorage(CONST.STORAGE.TABS_ACTIVE_KEY, action.payload);
     },
   });
 };
