+++ src/renderer/store/listeners/web.ts
@@ -1,13 +1,13 @@
 import listenerMiddleware from '@/store/listeners';
 import { syncWebConfigAction } from '@/store/features/web';
-import * as Utils from '@/utils';
+import * as Enhancement from '@/utils/enhancement';
 import * as CONST from '@/constants';
 
 export default () => {
   listenerMiddleware.startListening({
     actionCreator: syncWebConfigAction,
     effect: async (action) => {
-      Utils.SetStorage(CONST.STORAGE.WEB_SETTING, action.payload.webConfig);
+      Enhancement.SetStorage(CONST.STORAGE.WEB_SETTING, action.payload.webConfig);
     },
   });
 };
