+++ src/renderer/workers/index.ts
@@ -1,0 +1,1 @@
+export const sortWorker = new Worker(new URL('./sort.worker.ts', import.meta.url));
